Refactor calendar to its own file
